% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_utils.R
\name{cheops_cancel}
\alias{cheops_cancel}
\title{Cancel a running SLURM job}
\usage{
cheops_cancel(id)
}
\arguments{
\item{id}{the id of the job}
}
\value{
a table containing the currently running jobs
}
\description{
Cancel a running SLURM job
}
