% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{cheops_getlog}
\alias{cheops_getlog}
\title{Get a log file from a submitted job}
\usage{
cheops_getlog(jobname)
}
\arguments{
\item{jobname}{the name of the job}
}
\value{
a raw character vector containing the lines of the log file
}
\description{
Get a log file from a submitted job
}
