% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{cheops_install}
\alias{cheops_install}
\title{Install a package on the cluster}
\usage{
cheops_install(package, module = getOption("cheopsr.module"),
  lib = getOption("cheopsr.libloc"))
}
\arguments{
\item{package}{name of the package as a character string}

\item{module}{r module to be loaded}

\item{lib}{lib location on the cluster}
}
\value{
character vector of R console output
}
\description{
Install a package on the cluster
}
\details{
if the lib does not exist, a folder is created
}
