% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_utils.R
\name{cheops_jobs}
\alias{cheops_jobs}
\title{List the users running SLURM jobs}
\usage{
cheops_jobs()
}
\value{
a table containing the currently running jobs
}
\description{
List the users running SLURM jobs
}
