% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{cheops_modules}
\alias{cheops_modules}
\title{List all or R specific modules available on CHEOPS}
\usage{
cheops_modules(ronly = TRUE)
}
\arguments{
\item{ronly}{a logical value indicating if only R specifig modules should be returned}
}
\value{
a vector of module names
}
\description{
List all or R specific modules available on CHEOPS
}
