% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{cheops_readRDS}
\alias{cheops_readRDS}
\title{Restore a single R Object from the cluster}
\usage{
cheops_readRDS(file)
}
\arguments{
\item{file}{location of the .rds file on the cluster}
}
\value{
the file
}
\description{
Restore a single R Object from the cluster
}
