% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{cheops_readtable}
\alias{cheops_readtable}
\title{Read in a table file from the cluster}
\usage{
cheops_readtable(file, ...)
}
\arguments{
\item{file}{location of the .txt or .csv file on the cluster}

\item{...}{additional arguments passed to {\link[utils]{read.table}}}
}
\value{
a data.frame
}
\description{
Read in a table file from the cluster
}
