% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{cheops_run}
\alias{cheops_run}
\title{Run a function on the cluster}
\usage{
cheops_run(fun, options, jobname, args = list(), packages = c(),
  module = getOption("cheopsr.module"),
  account = getOption("cheopsr.account"), lib = getOption("cheopsr.libloc"))
}
\arguments{
\item{fun}{a function. The function must accept an argument with the name \code{cl}}

\item{options}{named list of cluster options generated by {\link[cheopsr]{cheops_slurmcontrol}}}

\item{args}{optionally a list of additional arguments passed to \code{fun} via {\link[base]{do.call}}}

\item{packages}{character vector of package names to be loaded on the cluster}

\item{module}{name of r module to be loaded}

\item{account}{name of the account to be charged}

\item{lib}{location of the r lib on the cluster}
}
\value{
a named list with the name and id of the job and the file path on the cluster for the results
}
\description{
Run a function on the cluster
}
\details{
starts a slurm job on the cluster.
after the job is terminated, results have to be retrieved manually with the command: \code{cheops_readRDS(<filepath>)}.
}
\seealso{
{\link[parallel]{parLapplyLB}}
}
