% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{cheops_saveRDS}
\alias{cheops_saveRDS}
\title{Save an object on the cluster}
\usage{
cheops_saveRDS(obj, file)
}
\arguments{
\item{obj}{R object}

\item{file}{location of the .rds file on the cluster}
}
\value{
nothing
}
\description{
Save an object on the cluster
}
