% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{cheops_slurmcontrol}
\alias{cheops_slurmcontrol}
\title{Options for a SLURM task}
\usage{
cheops_slurmcontrol(nodes, tasks, mem, time, partition = NULL)
}
\arguments{
\item{nodes}{Number of nodes between 1 - 128}

\item{tasks}{Number of tasks \strong{per} node}

\item{mem}{Memory per node as a string (e.g. "4gb")}

\item{time}{Wall time of the job in the format "hh:mm:ss"}
}
\value{
a named list of options
}
\description{
Generates a list of options to be used in {\link[cheopsr]{cheops_submit}}
}
