% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{cheops_submit}
\alias{cheops_submit}
\title{Submit a sbatch job to the cluster}
\usage{
cheops_submit(jobname, rscript, options, module = getOption("cheopsr.module"),
  account = getOption("cheopsr.account"), lib = getOption("cheopsr.libloc"))
}
\arguments{
\item{rscript}{location of the r script file to be executed}

\item{options}{named list of cluster options generated by {\link[cheopsr]{cheops_slurmcontrol}}}

\item{module}{name of r module to be loaded}

\item{account}{name of the account to be charged}

\item{lib}{location of the r lib on the cluster}

\item{name}{name of the job}
}
\value{
a numeric job id
}
\description{
Submit a sbatch job to the cluster
}
